[Feat] : SocketConfig 코드 추가
