--- src/main/java/com/daangn/dangunmarket/global/config/ChatConfig.java
@@ -1,28 +1,0 @@
-package com.daangn.dangunmarket.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-
-@Configuration
-@EnableWebSocketMessageBroker
-public class ChatConfig implements WebSocketMessageBrokerConfigurer {
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/ws/chat").setAllowedOriginPatterns("*").withSockJS();
-        registry.addEndpoint("/ws/chat").setAllowedOriginPatterns("*");
-    }
-
-    @Override
-    public void configureMessageBroker(MessageBrokerRegistry registry) {
-        // 메세지 구독 요청 url -> 메세지 받을 때
-        registry.enableSimpleBroker("/sub");
-
-        // 메세지 발행 요청 url -> 메세지 보낼 때
-        registry.setApplicationDestinationPrefixes("/pub");
-    }
-
-}
